rpcperms: allow some RPCs before `rpcActive` state
Change the `InterceptorChain` behavior to allow a remote signer to call
the `walletrpc.SignCoordinatorStreams` while the `rpcState` is set to
`allowRemoteSigner`. This state precedes the `rpcActive` state, which
allows all RPCs.

This change is necessary because `lnd` needs the remote signer to be
connected before some of the internal dependencies for RPC sub-servers
can be created. These dependencies must be inserted into the sub-servers
before moving the `rpcState` to `rpcActive`.