rpcwallet: add `RemoteSignerConnection` interface
RemoteSignerConnection is an interface that abstracts the communication
with a remote signer. It extends the RemoteSignerRequests interface, and
adds som additional methods to manage the connection and verify the
health of the remote signer.

As we'll add an outbound remote signer implementation in upcoming
commits, we need this interface to abstract the commonalities of both
the inbound (the current remote signer implementation) and outbound
remote signer implementations, so that the RPCKeyRing doesn't need to
know which type it's using.