rpcwallet: add `OutboundConnection` implementation
This commit wraps the connection to the current remote signer
implementation in the new `RemoteSignerConnection` interface within an
`OutboundConnection` struct.

The name `OutboundConnection` is chosen to reflect the fact that the
watch-only node makes an outbound connection to the inbound remote
signer in the current implementation, and this connection instance is
used on the watch-only node's side.