lncfg: allow `watchonlynode.enable`
As all the necessary pieces on the signer node side to let the remote
signer make an outbound connection to the watch-only node are in place,
the `lncfg` package can now permit the `watchonlynode.enable` setting to
be configured.

Note that we still haven't created the implementation on the watch-only
node side to accept the connection and send the requests over the
stream. This will be added in the upcoming commits.